Split lexer number scanning
