+++ src/lexer/mod.rs
@@ -1,3 +1,4 @@
+mod numbers;
 mod scan;
 mod string_interpolation;
 mod strings;
